range to llint

+++ header.h
@@ -4,7 +4,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <time.h>
-typedef long int li;
+typedef long long int li;
 #define true 1
 #define false 0
 #endif
